Making Authorized Requests

+++ frontend/src/Components/WorkoutDetails.jsx
@@ -2,12 +2,19 @@ import { DELETE_WORKOUT } from "../context/WorkoutContextTypes";
 import { useWorkoutContext } from "../Hooks/useWorkoutContext";
 
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
+import { useAuthContext } from "../Hooks/useAuthContext";
 
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutContext();
+  const { user } = useAuthContext();
+
   const handleClick = async () => {
+    if (!user) return;
     const response = await fetch("/api/workouts/" + workout._id, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
     });
     const json = await response.json();
